+++ cmdline/touch.c
@@ -137,7 +137,7 @@ void state_touch(struct snapraid_state* state)
 				/* state changed, we need to update it */
 				state->need_write = 1;
 
-				log_tag("touch:%s:%s: %" PRIu64 ".%d\n", disk->name, esc_tag(file->sub, esc_buffer), (uint64_t)st.st_mtime, (int)STAT_NSEC(&st));
+				log_tag("touch:%s:%s: %" PRIu64 ".%d\n", disk->name, esc_tag(file->sub, esc_buffer), (uint64_t)st.st_mtime, STAT_NSEC(&st));
 				msg_info("touch %s\n", fmt_term(disk, file->sub, esc_buffer));
 			}
 		}
